Extract GASP trajectory runtime seam
